pyproject.toml
[tool.poetry]
name = "web-backend"
version = "0.1.0"
description = ""
authors = ["namakemono-san"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.106.0"
uvicorn = {extras = ["standard"], version = "^0.25.0"}
aiohttp = "^3.8.6"
rethinkdb = "^2.4.10.post1"
python-dotenv = "^1.0.0"
apscheduler = "^3.10.4"
fastapi-utils = "^0.2.1"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
pyproject-flake8 = "^0.0.1-alpha.4"
isort = "^5.10.1"
poethepoet = "^0.13.1"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120

[tool.flake8]
max-line-length = 120
ignore = "E203,W503,C901"
max-complexity = 10
extend-exclude = ".venv"

[tool.poe.tasks]
lint = { sequence = ["lint:flake8", "lint:isort"] }
"lint:flake8" = "poetry run pflake8 ."
"lint:isort" = "isort -c ."
format = { sequence = ["format:black", "format:isort"] }
"format:black" = "poetry run black ."
"format:isort" = "poetry run isort ."

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"